requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.0
django-filter==24.3
django-import-export==4.1.1
django-jazzmin==3.0.0
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
